added a update Reviews component

+++ src/compontents/layout-components/WatchList.jsx
@@ -1,6 +1,7 @@
 import React, { useContext, useEffect, useState } from 'react';
 import Background from '../../assets/background2.jpg'
 import { AuthContext } from '../../provider/AuthProvider';
+import Swal from 'sweetalert2';
 
 const WatchList = () => {
 
@@ -16,13 +17,32 @@ const WatchList = () => {
     },[user]);
 
     const handleDelete = id =>{
-        fetch(`http://localhost:5000/watchList/${id}`,{
-            method:'DELETE',
-        })
-        .then(res => res.json())
-        .then(()=>{
-            setWatchList(watchList.filter(item => item._id !== id))
-        })
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+          }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/watchList/${id}`,{
+                    method:'DELETE',
+                })
+                .then(res => res.json())
+                .then(()=>{
+                    setWatchList(watchList.filter(item => item._id !== id));
+                    Swal.fire({
+                        title: "Deleted!",
+                        text: "Your file has been deleted.",
+                        icon: "success"
+                      });
+                })
+             
+            }
+          });
+      
     }
     return (
        <div style={{
